pyproject.toml
[project]
name = "proyecto"
version = "0.1.0"
description = "Pipeline de datos de restaurantes europeos (TripAdvisor)"
requires-python = ">=3.10"
dependencies = [
    "apache-airflow>=2.9.0",
    "pandas>=2.0.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "kafka-python>=2.0.0",
    "folium>=0.17.0",
]

[tool.uv]
dev-dependencies = []
